requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
dj-database-url==2.3.0
Django==5.2
django-widget-tweaks==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
ruff==0.11.7
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
